refactor: reduce nested queries in ClaimRewards and RescueTokens
